fix: modified the fee input unit to xrp from drops

+++ packages/cysync-core/src/dialogs/Send/Dialogs/Components/FeeSection/XrpInput.tsx
@@ -24,7 +24,7 @@ export const XrpInput: React.FC<XrpInputProps> = ({
       value={value.toString()}
       postfixText={unit}
       onChange={handleChange}
-      valueType="integer"
+      valueType="float"
     />
   );
 };
